1.Flux debug version, first commit

+++ Assets/Code/Game@hotfix/demo_Manager_AutoRegister_And_Event/Event_demo1.cs
@@ -1,7 +1,7 @@
 namespace Code.Game.demo_Manager_AutoRegister_And_Event
 {
     
-    [DemoEvent(DemoEventEnum.TestEvent2)]
+    [DemoEvent((int)DemoEventEnum.TestEvent2)]
     public class Event_demo1:IDemoEvent
     {
         public void Do()
