+++ Assets/Code/Game@hotfix/demo_Manager_AutoRegister_And_Event/Event_demo2.cs
@@ -4,7 +4,7 @@ namespace Code.Game.demo_Manager_AutoRegister_And_Event
     /// <summary>
     /// 标签一定要加！！！！！！
     /// </summary>
-    [DemoEvent(DemoEventEnum.TestEvent2)]
+    [DemoEvent((int)DemoEventEnum.TestEvent2)]
     public class Event_demo2:IDemoEvent
     {
         public void Do()
